requirements.txt
amqp==2.1.4
appdirs==1.4.0
billiard==3.5.0.2
boto3==1.4.4
botocore==1.5.10
celery==4.0.2
decorator==4.0.11
Django==1.10.5
django-extensions==1.7.6
django-storages==1.5.2
djangorestframework==3.5.4
djangorestframework-jwt==1.9.0
docutils==0.13.1
drfdocs==0.0.11
ipython==5.2.2
ipython-genutils==0.1.0
jmespath==0.9.1
kombu==4.0.2
mysqlclient==1.3.9
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.13
ptyprocess==0.5.1
py==1.4.32
Pygments==2.2.0
PyJWT==1.4.2
pyparsing==2.1.10
pytest==3.0.6
pytest-django==3.1.2
python-dateutil==2.6.0
pytz==2016.10
redis==2.10.5
requests==2.13.0
s3transfer==0.1.10
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
vine==1.1.3
wcwidth==0.1.7
Werkzeug==0.11.15
